Create Chapter_7_files
Python for Informatics

Exercise 7.2

Write a program to prompt for a ﬁle name, and then read through the ﬁle and look for lines of the form:
X-DSPAM-Confidence: 0.8475
When you encounter a line that starts with “X-DSPAM-Conﬁdence:” pull apart the line to extract the ﬂoating-point number on the line. Count these lines and then compute the total of the spam conﬁdence values from these lines. When you reach the end of the ﬁle, print out the average spam conﬁdence.
Enter the file name: mbox.txt Average spam confidence: 0.894128046745
Enter the file name: mbox-short.txt Average spam confidence: 0.750718518519
Test your ﬁle on the mbox.txt and mbox-short.txt ﬁles